Use allocated workspace dir instead of hardcoded
When a job runs in a node it allocates a workspace for the execution
to avoid running in the incorrect workspace in case of multiple jobs
we were hardcoding the directory where we wanted to run. This commit
changes that so we run in the workspace allocated for the node but
we save that location in an environment variable to make sure we
run in the same location during the rest of the workflow.