Fix the upgrade playbook to set the env correctly
The upgrade_osa playbook needs the environment variable
I_REALLY_KNOW_WHAT_I_AM_DOING set to true in order to work. This
was being provided in the playbook but incorrectly, this commit
fixes the way the env variable is being set up. It also receives
the release version as variable instead of being hardcoded so it
can be used to upgrade to different release version.