Adding the base tempest configuration to osic repo
Jenkins is using a tempest.conf file that includes common
values used by OSA as a base for the tempest configuration file,
this base config file was being downloaded from a repo different
than the OSIC one. This commit includes that base config file into
the OSIC repo and changes the code to download this file from
this repo instead.