Renamed the pipeline to something more meaningful
The onMetal Simple Upgrade pipeline was just called Jenkinsfile
which is a convention name for pipelines, so it would make sense
to keep it like that if there was only one pipeline. Since we
want to have multiple pipelines there is a need for renaming it to
comething that makes sense.