Adding a function to install ElasticSearch parse
The elasticSearch parser was being installed as part of the function
that the pipeline was using for parsing the results. The parser
installation was removed from the function that uses the parser in a
previous commit. This commit adds a separate function to install the
parser and or upgrade it in the system.