Adding a cleanup function for test results
Usually when the CI runs it creates a whole new system, so all
directories will be new, and no cleanup is necessary for most of
the things at the end of the pipeline. However once in a while we
want to re-run some tests in an existing environment, in those cases
if there are directories with test results they could be mixed with
the new results. This commit cleans up those directories if they
exist.