Adding a wait time for the output file to exist
When the api uptime tests run they run in the background, when
the command to stop them is run, the script will collect the results
and will copy them to an output file. This can take some fractions
of a second, if the pipeline tries to run during that time it will
not find the results. This commit makes sure the file is out before
continueing with the pipeline.